% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequestList.r
\name{addRequest,requestList,list-method}
\alias{addRequest,requestList,list-method}
\title{adding request to list}
\usage{
\S4method{addRequest}{requestList,list}(r, entry)
}
\arguments{
\item{r}{\itemize{
\item the request list
}}

\item{entry}{\itemize{
\item the new request
}}
}
\description{
adding request to list
}
