% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequestList.r
\name{addRequest}
\alias{addRequest}
\title{adding request to list}
\usage{
addRequest(r, entry)
}
\arguments{
\item{r}{\itemize{
\item the request list
}}

\item{entry}{\itemize{
\item the new request
}}
}
\description{
adding request to list
}
