% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSLTR.r
\name{addSLTR}
\alias{addSLTR}
\title{add SLTR to selectedSamples}
\usage{
addSLTR(ssl)
}
\arguments{
\item{ssl}{\itemize{
\item selectedSamples
}}
}
\value{
selectedSamples with SLTR
}
\description{
add SLTR to selectedSamples
}
