% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{API base url and endpoint setup}
\usage{
api(x)
}
\arguments{
\item{x}{A URL}
}
\description{
API base url and endpoint setup
}
\examples{
\dontrun{
# Set base url
## works with full or partial URLs
api('https://api.github.com/')
api('http://api.gbif.org/v1')
api('api.gbif.org/v1')

## works with ports, full or partial
api('http://localhost:9200')
api('localhost:9200')
api(':9200')
api('9200')
api('9200/stuff')

# set paths
## NSE
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, issues)
## SE
api('https://api.github.com/') \%>\%
  api_path_('repos', 'ropensci', 'rgbif', 'issues')

# template
repo_info <- list(username = 'craigcitro', repo = 'r-travis')
api('https://api.github.com/') \%>\%
  api_template(template = 'repos/{{username}}/{{repo}}/issues', data = repo_info)
}
}
\seealso{
Other dsl: 
\code{\link{api_body}()},
\code{\link{api_config}()},
\code{\link{api_error_handler}()},
\code{\link{api_path}()},
\code{\link{api_query}()},
\code{\link{api_template}()},
\code{\link{auth}}
}
\concept{dsl}
