% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/body.R
\name{api_body}
\alias{api_body}
\alias{api_body_}
\title{Query construction}
\usage{
api_body(.data, ..., body_value = NULL)

api_body_(.data, ..., .dots, body_value)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Query construction
}
\examples{
## NSE
dd <- api("http://httpbin.org/post")
dd \%>\% api_body(body_value = FALSE)
dd \%>\% api_body(body_value = "NULL")
dd \%>\% api_body(body_value = "")

# upload a file
file <- "~/httsnap_test.txt"
cat("hello, world", file = file)
dd \%>\% api_body(body_value = upload_file("~/httsnap_test.txt"))

# A named list
dd \%>\% api_body(x = "A simple text string", y = "stuff")

## SE
api("http://api.plos.org/search") \%>\%
  api_query_(q = "ecology", wt = "json", fl = 'id', fl = 'journal')

\dontrun{
## NSE
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = 'id,journal') \%>\%
  Get()
## SE
api("http://api.plos.org/search") \%>\%
  api_query_(q = "ecology", wt = "json", fl = 'id', fl = 'journal') \%>\%
  Get()
}
}
\seealso{
Other dsl: \code{\link{api_config}};
  \code{\link{api_error_handler}};
  \code{\link{api_oauth1}}, \code{\link{api_oauth2}},
  \code{\link{api_simple_auth}}, \code{\link{auth}};
  \code{\link{api_query}}, \code{\link{api_query_}};
  \code{\link{api}}, \code{\link{api_path}},
  \code{\link{api_path_}}, \code{\link{api_template}}
}

