% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{api_cache}
\alias{api_cache}
\title{Caching helper}
\usage{
api_cache(.data, dir = NULL, ...)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{dir}{(character) Directory to cache in. Uses
\code{rappdirs::user_cache_dir()} by default}

\item{...}{ignored}
}
\description{
Caching helper
}
\examples{
\dontrun{
# cache
## first call is slower
api('http://localhost:5000') \%>\%
  api_path(get) \%>\%
  api_query(foo = "bar") \%>\%
  api_cache()

## second call is faster, pulling from cache
api('http://localhost:5000') \%>\%
  api_path(get) \%>\%
  api_query(foo = "bar") \%>\%
  api_cache()

# other egs
x <- api('api.crossref.org') \%>\%
  api_path(works) \%>\%
  api_query(rows = 1000) \%>\%
  api_cache()
}
}
