% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors.R
\name{api_error_handler}
\alias{api_error_handler}
\title{Error handler}
\usage{
api_error_handler(.data, fun)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{fun}{A function, either defined in the session, or a function available in loaded
or name-spaced packges}
}
\description{
Error handler
}
\examples{
# Use functions from httr
api('https://api.github.com/') \%>\%
 api_error_handler(stop_for_status)

api('https://api.github.com/') \%>\%
 api_error_handler(warn_for_status)

# Custom error handling functions
my_stop <- function(x) {
  if (x$status > 200) {
     warning("nope, try again", call. = FALSE)
  }
}
api("http://httpbin.org/status/404") \%>\%
 api_error_handler(my_stop)

\dontrun{
# Error handler
api("http://httpbin.org/status/404") \%>\%
 api_error_handler(stop_for_status) \%>\%
 Get()

# Custom error handling function
api("http://httpbin.org/status/404") \%>\%
 api_error_handler(my_stop) \%>\%
 Get()
}
}
\seealso{
Other dsl: \code{\link{api_body}}, \code{\link{api_body_}};
  \code{\link{api_config}}; \code{\link{api_oauth1}},
  \code{\link{api_oauth2}}, \code{\link{api_simple_auth}},
  \code{\link{auth}}; \code{\link{api_paging}};
  \code{\link{api_query}}, \code{\link{api_query_}};
  \code{\link{api}}, \code{\link{api_path}},
  \code{\link{api_path_}}, \code{\link{api_template}}
}

