% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_path.R
\name{api_path}
\alias{api_path}
\alias{api_path_}
\title{API paths}
\usage{
api_path(.data, ..., .dots)

api_path_(.data, ..., .dots)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
API paths
}
\examples{
\dontrun{
# set paths
## NSE
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, issues)
## SE
api('https://api.github.com/') \%>\%
  api_path_('repos', 'ropensci', 'rgbif', 'issues')
}
}
\seealso{
Other dsl: 
\code{\link{api_body}()},
\code{\link{api_config}()},
\code{\link{api_error_handler}()},
\code{\link{api_query}()},
\code{\link{api_template}()},
\code{\link{api}()},
\code{\link{auth}}
}
\concept{dsl}
