% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paging.R, R/query.R
\name{api_query}
\alias{api_query}
\alias{api_query_}
\title{Paging helpers}
\usage{
api_query(.data, ...)

api_query_(.data, ..., .dots)

api_query(.data, ...)

api_query_(.data, ..., .dots)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{.data}{Result of a call to \code{api}}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Paging helpers

Query construction
}
\examples{
## NSE
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = 'id,journal')
### Or
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = id, fl = journal)
## SE
api("http://api.plos.org/search") \%>\%
  api_query_(q = "ecology", wt = "json", fl = 'id', fl = 'journal')

\dontrun{
## NSE
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = 'id,journal') \%>\%
  Get()
## SE
api("http://api.plos.org/search") \%>\%
  api_query_(q = "ecology", wt = "json", fl = 'id', fl = 'journal') \%>\%
  Get()
}
## NSE
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = 'id,journal')
### Or
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = id, fl = journal)
## SE
api("http://api.plos.org/search") \%>\%
  api_query_(q = "ecology", wt = "json", fl = 'id', fl = 'journal')

\dontrun{
## NSE
api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = 'id,journal') \%>\%
  Get()
## SE
api("http://api.plos.org/search") \%>\%
  api_query_(q = "ecology", wt = "json", fl = 'id', fl = 'journal') \%>\%
  Get()
}
}
\seealso{
Other dsl: \code{\link{api_body}}, \code{\link{api_body_}};
  \code{\link{api_config}};
  \code{\link{api_error_handler}};
  \code{\link{api_oauth1}}, \code{\link{api_oauth2}},
  \code{\link{api_simple_auth}}, \code{\link{auth}};
  \code{\link{api}}, \code{\link{api_path}},
  \code{\link{api_path_}}, \code{\link{api_template}}

Other dsl: \code{\link{api_body}}, \code{\link{api_body_}};
  \code{\link{api_config}};
  \code{\link{api_error_handler}};
  \code{\link{api_oauth1}}, \code{\link{api_oauth2}},
  \code{\link{api_simple_auth}}, \code{\link{auth}};
  \code{\link{api}}, \code{\link{api_path}},
  \code{\link{api_path_}}, \code{\link{api_template}}
}

