% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_retry.R
\name{api_retry}
\alias{api_retry}
\title{Retry on failed request N times}
\usage{
api_retry(.data, n, time = 1)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{n}{Number of times to repeat the request}

\item{time}{Number of seconds to delay between repeated calls}
}
\description{
FIXME: still need to implement doing this in the request
}
\details{
This doesn't use the retry option within curl itself,
as it's not available via the \code{curl} R client. Instead, we
retry X times you specify, if the previous call failed.
}
\examples{
\dontrun{
api('http://127.0.0.1:8080') \%>\%
   api_retry(n = 5)

api('http://127.0.0.1:8080') \%>\%
   api_retry(n = 5, time = 2)
}
}

