% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_template.R
\name{api_template}
\alias{api_template}
\title{API path template}
\usage{
api_template(.data, template, data)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{template}{Template to contstruct API route}

\item{data}{Data to pass to the template parameter}
}
\description{
API path template
}
\examples{
\dontrun{
repo_info <- list(username = 'craigcitro', repo = 'r-travis')
api('https://api.github.com/') \%>\%
  api_template(template = 'repos/{{username}}/{{repo}}/issues', data = repo_info)
}
}
\seealso{
Other dsl: \code{\link{api_body}},
  \code{\link{api_config}},
  \code{\link{api_error_handler}}, \code{\link{api_path}},
  \code{\link{api_query}}, \code{\link{api}},
  \code{\link{auth}}
}

