% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_write.R
\name{api_write}
\alias{api_write}
\title{Write helper}
\usage{
api_write(.data, file, overwrite = FALSE, ...)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{file}{(character) Full file path to write to}

\item{overwrite}{(logical) Will only overwrite existing path if \code{TRUE}}

\item{...}{ignored for now}
}
\description{
Write helper
}
\examples{
\dontrun{
## write to disk
ff <- tempfile(fileext = ".json")
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, issues) \%>\%
  api_write(ff)
jsonlite::fromJSON(ff)
}
}

