% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df}
\alias{as_df}
\title{Attempt to convert list to a data.frame}
\usage{
as_df(x, clean = TRUE)
}
\arguments{
\item{x}{input, a list}

\item{clean}{(logical) clean 0 length elements. Default: \code{TRUE}}
}
\description{
Attempt to convert list to a data.frame
}
\examples{
mtlist <- apply(mtcars, 1, as.list)
as_df(mtlist)

mtlist <- apply(mtcars, 1, as.list)
mtlist[[1]]$mpg <- list()
as_df(mtlist)
}

