% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNames.r
\name{cleanNames}
\alias{cleanNames}
\title{cleaning names for automation (no underscore is allowed as
it is a separator in long names)}
\usage{
cleanNames(names)
}
\arguments{
\item{names}{\itemize{
\item a name or an array of
}}
}
\value{
clean name(s)
}
\description{
cleaning names for automation (no underscore is allowed as
it is a separator in long names)
}
\examples{
originalID <- c("ddd.aaa", "ddd uuu", "ddd+aaa", "ddd*yyy", "ddd#dd", "ddd_fff")
originalID <-c(originalID, "ddd$ddd", "ddd@ddd", "dd_aa", "dad*")
cleanNames(originalID)
}
