% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\name{fillRequest,request,list-method}
\alias{fillRequest,request,list-method}
\title{filling requests}
\usage{
\S4method{fillRequest}{request,list}(r, request)
}
\arguments{
\item{r}{the request}

\item{request}{the parameters to fill}
}
\description{
filling requests
}
