% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\name{fillRequest}
\alias{fillRequest}
\title{filling requests}
\usage{
fillRequest(r, request)
}
\arguments{
\item{r}{the request}

\item{request}{the parameters to fill}
}
\description{
filling requests
}
