% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{findEmptyPositions}
\alias{findEmptyPositions}
\title{find empty positions}
\usage{
findEmptyPositions(positions, boxDim = c(8, 10), by = "col", exclude = c())
}
\arguments{
\item{positions}{\itemize{
\item the position or an array of positions
}}

\item{boxDim}{\itemize{
\item the dimension of the box (row, column)
}}

\item{by}{\itemize{
\item the margin
}}
}
\value{
empty positions
}
\description{
find empty positions
}
