% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.R
\name{Get}
\alias{Get}
\title{Get a url, with sensible defaults}
\usage{
Get(.data, parse = TRUE, ...)
}
\arguments{
\item{.data}{A request object}

\item{parse}{(logical) Attempt to parse data to data.frame if possible. Default: TRUE}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}.}
}
\description{
Get a url, with sensible defaults
}
\details{
Attempts to simplify the http request process by using sensible defaults:
\itemize{
 \item GET by default: you most likely want to use \code{\link[httr]{GET}}
 \item You most likely want a data.frame back, so we attempt to coerce to a data.frame
}
}
\examples{
\dontrun{
"https://api.github.com/" \%>\%
   Get()

"https://api.github.com/" \%>\%
   Progress() \%>\%
   Verbose() \%>\%
   Get()

"https://api.github.com/" \%>\%
   Timeout(3) \%>\%
   Get()

"http://api.crossref.org/works/" \%>\%
   User_agent("howdydoodie") \%>\%
   Get()

"http://api.plos.org/search" \%>\%
   Query(q = "*:*", wt = "json") \%>\%
   Get() \%>\%
   .$response \%>\%
   .$docs
}
}

