% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDevice.r
\name{getDevice}
\alias{getDevice}
\title{retrieve device name from ID using googlesheet}
\usage{
getDevice(what)
}
\arguments{
\item{what}{\itemize{
\item what to retrive c(deviceID = 1)
}}
}
\value{
the name of the device
}
\description{
retrieve device name from ID using googlesheet
}
