% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{getPlatePos}
\alias{getPlatePos}
\title{create array of positions}
\usage{
getPlatePos(boxDim = c(8, 12), by = "row")
}
\arguments{
\item{boxDim}{\itemize{
\item the dimension of the box (row, column)
}}

\item{by}{\itemize{
\item the direction of filling the box
}}
}
\value{
a data.frame with row and column positions
}
\description{
create array of positions
}
\examples{
g <- getPlatePos(by = "col")
}
