% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{Authenticate}
\alias{Body}
\alias{Progress}
\alias{Query}
\alias{Timeout}
\alias{User_agent}
\alias{Verbose}
\alias{helpers}
\title{http request helpers}
\usage{
Progress(.data)

Verbose(.data)

Timeout(.data, seconds)

User_agent(.data, agent)

Authenticate(.data, user, password, type = "basic")

Query(.data, ...)

Body(.data, ...)
}
\arguments{
\item{.data}{A request object}

\item{seconds}{Number of seconds to wait for a response until giving up. Can
not be less than 1 ms (aka: 0.001 seconds).}

\item{agent}{An agent character string}

\item{user}{A user name}

\item{password}{A password}

\item{type}{Type of authenticaion. Default: basic}

\item{...}{Further args}
}
\description{
Various helpers as functions that allow chaining together options to
pass to e.g., \code{\link{Get}}
}

