% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_iterator.R
\name{http}
\alias{http}
\alias{http_client}
\title{Make a HTTP request}
\usage{
http(req)

http_client(req)
}
\arguments{
\item{req}{A \code{req} class object}
}
\description{
Make a HTTP request
}
\details{
By default, a GET request is made.
}
\examples{
\dontrun{
# high level - http()
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, commits) \%>\%
  http()

# low level - http_client()
res <- api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, commits) \%>\%
  api_paging(limit = 220, limit_max = 100) \%>\%
  http_client()
res$count()
res$result
res$links
res$parse()


# with paging
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, commits) \%>\%
  api_paging(limit = 220, limit_max = 100) \%>\%
  http()
}
}

