% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/httsnap-package.R
\docType{package}
\name{httsnap-package}
\alias{httsnap}
\alias{httsnap-package}
\title{Easy http}
\description{
Easy http
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

