% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\name{initialize,request-method}
\alias{initialize,request-method}
\title{initialize request}
\usage{
\S4method{initialize}{request}(.Object)
}
\arguments{
\item{.Object}{the request object}
}
\description{
initialize request
}
