% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{numToPos}
\alias{numToPos}
\title{convert a numeric position to a letter position}
\usage{
numToPos(num, boxDim = c(8, 12))
}
\arguments{
\item{num}{\itemize{
\item the position or an array of positions
}}

\item{boxDim}{\itemize{
\item the dimension of the box (row, column)
}}
}
\value{
the letter position or an array of them
}
\description{
convert a numeric position to a letter position
}
