% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peep.R
\name{peep}
\alias{peep}
\title{Peek at a query}
\usage{
peep(.data)
}
\arguments{
\item{.data}{(list) input, using higher level interface}
}
\description{
Peek at a query
}
\examples{
\dontrun{
api('https://api.github.com/') \%>\% peep
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, issues) \%>\%
  peep

repo_info <- list(username = 'craigcitro', repo = 'r-travis')
api('https://api.github.com/') \%>\%
  api_template(template = 'repos/{{username}}/{{repo}}/issues', data = repo_info) \%>\%
  peep

api("http://api.plos.org/search") \%>\%
  api_query(q = ecology, wt = json, fl = id, fl = journal) \%>\%
  peep
}
}

