% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{posToNum}
\alias{posToNum}
\title{convert position to numeric position}
\usage{
posToNum(pos, boxDim = c(8, 12))
}
\arguments{
\item{pos}{\itemize{
\item the position or an array of positions
}}

\item{boxDim}{\itemize{
\item the dimension of the box (row, column)
}}
}
\value{
an array of numeric positions
}
\description{
convert position to numeric position
}
\examples{
posToNum(c("B2", "C3"), boxDim = c(10,10))
}
