% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLayout.r
\name{printLayout}
\alias{printLayout}
\title{print html layout of plates}
\usage{
printLayout(selectedSamples, boxDim = c(8, 10), by = "row", chk = FALSE)
}
\arguments{
\item{selectedSamples}{\itemize{
\item the list of sample to process
}}
}
\value{
an html table
}
\description{
print html layout of plates
}
