% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequestList.r
\name{printRequest,requestList,list-method}
\alias{printRequest,requestList,list-method}
\title{printing a request list}
\usage{
\S4method{printRequest}{requestList,list}(r, options)
}
\arguments{
\item{r}{\itemize{
\item the request list
}}

\item{options}{\itemize{
\item options
}}
}
\description{
printing a request list
}
