% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequestList.r
\name{printRequest}
\alias{printRequest}
\title{printing a request list}
\usage{
printRequest(r, options)
}
\arguments{
\item{r}{\itemize{
\item the request list
}}

\item{options}{\itemize{
\item options
}}
}
\description{
printing a request list
}
