% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{rate_limit}
\alias{rate_limit}
\title{Rate limiting}
\usage{
rate_limit(.data, value, period, on_limit = with_stop())
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{value}{(integer) Value of rate limit, number of requests allowed}

\item{period}{Time period, e.g., 1 min, 60 min, 1 hr, 24 hrs}

\item{on_limit}{What to do on reaching rate limit. See Details.}
}
\description{
FIXME: still need to implement doing this in the request
}
\details{
\code{on_limit} options:
\itemize{
 \item stop - and give error message to use
 \item warn - and give error message to use
 \item wait - and give max time to wait
}
}
\examples{
\dontrun{
qr <- api('https://api.github.com/') \%>\%
 api_path(repos, ropensci, rgbif, issues)

qr \%>\% rate_limit(value = 5, period = "24 hrs")
qr \%>\% rate_limit(value = 5000, period = "24 hrs")
qr \%>\% rate_limit(value = 10, period = "5 min")
qr \%>\% rate_limit(value = 10, period = "5 min", on_limit = with_wait(1))
}
}

