% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{raw}
\alias{raw}
\alias{raw_}
\title{get raw data}
\usage{
raw(.obj = list(), ...)

raw_(.data = list(), ..., .dots, parse = TRUE)
}
\arguments{
\item{.obj}{Data}

\item{...}{further args}

\item{.data}{A request object}

\item{.dots}{xxx}

\item{parse}{(logical) To parse or not}
}
\description{
get raw data
}

