% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{RCToPos}
\alias{RCToPos}
\title{convert row and column to a position}
\usage{
RCToPos(row, col)
}
\arguments{
\item{row}{\itemize{
\item the row or an array of rows
}}

\item{col}{\itemize{
\item the column or an array of columns
}}
}
\value{
the position or an array of positions
}
\description{
convert row and column to a position
}
