% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\docType{class}
\name{request-class}
\alias{request-class}
\title{An S4 class for request}
\value{
a dataElement
}
\description{
An S4 class for request
}
\section{Slots}{

\describe{
\item{\code{.Data}}{the data matrix}

\item{\code{varName}}{a vector containing the name of each variable}

\item{\code{obsDescr}}{a data.frame containing experimental
conditions and a field called sampleID that MUST be unique.}

\item{\code{type}}{type can be NMR, MSU, MST, ANN}
}}

\examples{

new("request")

}
