% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-package.R
\docType{package}
\name{request-package}
\alias{request}
\alias{request-package}
\title{request: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jwist/request}
  \item Report bugs at \url{https://github.com/jwist/request/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
