% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry}
\usage{
retry(.data, times)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{times}{(integer) Number of times to retry the request}
}
\description{
FIXME: still need to implement doing this in the request
}
\details{
This doesn't use the retry option within curl itself,
as it's not available via the \code{curl} R client. Instead, we
retry X times you specify, if the previous call failed.
}
\examples{
\dontrun{
api('https://api.github.com/') \%>\%
 api_path(repos, ropensci, rgbif, issues) \%>\%
 retry(times = 5)
}
}

