% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRun.r
\name{runMS_LIPIDS}
\alias{runMS_LIPIDS}
\title{write run file for MS-LIPIDS}
\usage{
runMS_LIPIDS(
  selectedSamples,
  runName,
  projectName,
  matrixID,
  deviceID,
  methodID,
  LTR_NAME,
  date
)
}
\arguments{
\item{selectedSamples}{\itemize{
\item the selected samples for run
}}

\item{runName}{\itemize{
\item the name of the run
}}

\item{projectName}{\itemize{
\item the name of the project
}}

\item{matrixID}{\itemize{
\item the id of the sample matrix
}}

\item{deviceID}{\itemize{
\item the id of the device
}}

\item{methodID}{\itemize{
\item the id of the method
}}

\item{date}{\itemize{
\item the date of the run
}}
}
\value{
void
}
\description{
write run file for MS-LIPIDS
}
