% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRun.r
\name{runMS_TRYE}
\alias{runMS_TRYE}
\title{write run file for MS-TRY}
\usage{
runMS_TRYE(
  selectedSamples,
  runName,
  projectName,
  matrixName,
  deviceName,
  methodName,
  LTR_NAME,
  date
)
}
\arguments{
\item{selectedSamples}{\itemize{
\item the selected samples for run
}}

\item{runName}{\itemize{
\item the name of the run
}}

\item{projectName}{\itemize{
\item the name of the project
}}

\item{matrixName}{\itemize{
\item the id of the sample matrix
}}

\item{deviceName}{\itemize{
\item the id of the device
}}

\item{methodName}{\itemize{
\item the id of the method
}}

\item{LTR_NAME}{\itemize{
\item the name of LTR sample
}}

\item{date}{\itemize{
\item the date of the run
}}
}
\value{
void
}
\description{
write run file for MS-TRY
}
