% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRun.r
\name{runNMR}
\alias{runNMR}
\title{write run file for NMR}
\usage{
runNMR(selectedSamples, matrixID, deviceID, methodID, cohortName, projectName)
}
\arguments{
\item{selectedSamples}{\itemize{
\item the selected samples for run
}}

\item{matrixID}{\itemize{
\item the id of the samples matrix
}}

\item{deviceID}{\itemize{
\item the id of the device to run the experiments
}}

\item{methodID}{\itemize{
\item the id of the method to be run
}}

\item{cohortName}{\itemize{
\item the name of the cohort to run
}}

\item{projectName}{\itemize{
\item the name of the project
}}
}
\value{
write the file
}
\description{
write run file for NMR
}
