% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRun.r
\name{saveRun}
\alias{saveRun}
\title{write run file}
\usage{
saveRun(run, runName, methodName)
}
\arguments{
\item{run}{\itemize{
\item the selected samples for run
}}

\item{runName}{\itemize{
\item the run name
}}
}
\value{
write the file
}
\description{
write run file
}
