% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\name{setPosition,request,character-method}
\alias{setPosition,request,character-method}
\title{set position of requests}
\usage{
\S4method{setPosition}{request,character}(r, position)
}
\arguments{
\item{r}{the request}

\item{position}{the parameters to fill}
}
\description{
set position of requests
}
