% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\name{setPosition}
\alias{setPosition}
\title{set position of requests}
\usage{
setPosition(r, position)
}
\arguments{
\item{r}{the request}

\item{position}{the parameters to fill}
}
\description{
set position of requests
}
