% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{api_config}
\alias{api_config}
\title{Curl settings}
\usage{
api_config(.data, ...)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{...}{Comma separated list of unquoted variable names}
}
\description{
Curl settings
}
\examples{
# Config handler
api('https://api.github.com/') \%>\%
 api_config(verbose())

\dontrun{
# Full examples
api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, issues) \%>\%
  api_config(verbose()) \%>\%
  Get()
}
}
\seealso{
Other dsl: \code{\link{api_body}},
  \code{\link{api_error_handler}},
  \code{\link{api_paging}}, \code{\link{api_query}},
  \code{\link{api}}, \code{\link{auth}}
}

