% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{api_headers}
\alias{api_headers}
\alias{api_headers_}
\title{Headers}
\usage{
api_headers(.data, ..., .dots)

api_headers_(.data, ..., .dots)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{...}{Key value pairs of headers}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Headers
}
\examples{
\dontrun{
api('https://api.github.com/') \%>\%
   api_config(verbose()) \%>\%
   api_headers(`X-FARGO-SEASON` = 3) \%>\%
   peep

api('http://httpbin.org/headers') \%>\%
   api_headers(`X-FARGO-SEASON` = 3, `X-NARCOS-SEASON` = 5)
}
}

