% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paging.R
\name{api_paging}
\alias{api_paging}
\title{Paging helpers}
\usage{
api_paging(.data, limit, limit_max, offset = 0, by = NULL)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{limit}{Maximum number results desired.}

\item{limit_max}{Maximum number results allowed in each request.}

\item{offset}{Record to start at}

\item{by}{Chunk size, if chunking desired. Default:}
}
\description{
Paging helpers
}
\examples{
quer <- api('https://api.github.com/') \%>\%
  api_path(repos, ropensci, rgbif, issues) \%>\%
  api_query(state = open)

# per_page & page, w/ known max_limit
quer \%>\%
  api_paging(limit = 220, limit_max = 100)

##### Not working yet
# per_page & page
# quer \%>\%
#   api_paging(per_page = 10, page = 2)

# limit & offset
# quer \%>\%
#   api_paging(limit = 10, offset = 20)

# rows & start
# quer \%>\%
#   api_paging(rows = 10, start = 5)
}
\seealso{
Other dsl: \code{\link{api_body}}, \code{\link{api_body_}};
  \code{\link{api_config}};
  \code{\link{api_error_handler}};
  \code{\link{api_oauth1}}, \code{\link{api_oauth2}},
  \code{\link{api_simple_auth}}, \code{\link{auth}};
  \code{\link{api_query}}, \code{\link{api_query_}};
  \code{\link{api}}, \code{\link{api_path}},
  \code{\link{api_path_}}, \code{\link{api_template}}
}

