% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_write.R
\name{api_write}
\alias{api_write}
\title{Write helper}
\usage{
api_write(.data, file, overwrite = FALSE, ...)
}
\arguments{
\item{.data}{Result of a call to \code{api}}

\item{file}{(character) Full file path to write to}

\item{overwrite}{(logical) Will only overwrite existing path if \code{TRUE}}
}
\description{
Write helper
}
\examples{
## write to disk
ff <- tempfile(fileext = ".json")
api("http://127.0.0.1:8080") \%>\%
  api_path(repos) \%>\%
  api_write(ff) \%>\%
  http
jsonlite::fromJSON(ff)
}

