% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_iterator.R
\docType{data}
\name{GetIter}
\alias{GetIter}
\title{GET iterator}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x106f98e08> 
 - attr(*, "name")= chr "GetIter_generator"
}}
\usage{
GetIter
}
\description{
GET iterator
}
\keyword{datasets}

