% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{posToRC}
\alias{posToRC}
\title{convert A1 position to row and column}
\usage{
posToRC(pos, collapse = FALSE)
}
\arguments{
\item{pos}{\itemize{
\item an array of positions
}}

\item{collapse}{\itemize{
\item if TRUE, will collapse the output into a single string
}}
}
\value{
a data.frame with row and column positions
}
\description{
convert A1 position to row and column
}
