% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionUtils.r
\name{RCToNum}
\alias{RCToNum}
\title{convert row and column to a numeric position}
\usage{
RCToNum(row, col, boxDim = c(8, 12))
}
\arguments{
\item{row}{\itemize{
\item the row or an array of rows
}}

\item{col}{\itemize{
\item the column or an array of columns
}}

\item{boxDim}{\itemize{
\item the dimension of the box (row, column)
}}
}
\value{
the numeric position or an array of them
}
\description{
convert row and column to a numeric position
}
