% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classRequest.r
\docType{class}
\name{request-class}
\alias{request-class}
\title{An S4 class for request}
\value{
a object with request
}
\description{
An S4 class for request
}
\section{Slots}{

\describe{
\item{\code{sampleID}}{sampleID}

\item{\code{sampleType}}{sample type}

\item{\code{matrixID}}{id of sample matrix}

\item{\code{methodID}}{id of the method}

\item{\code{deviceID}}{id of the device}

\item{\code{projectName}}{name of the project}

\item{\code{cohortName}}{name of the cohort}

\item{\code{platePosition}}{position of the plate}

\item{\code{row}}{row of the sample}

\item{\code{column}}{column of the sample}

\item{\code{options}}{options}
}}

\examples{

new("request")

}
