% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-package.R
\docType{package}
\name{request-package}
\alias{request}
\alias{request-package}
\title{Easy http}
\description{
Easy http
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

